+++ src/media-utils.ts
@@ -2,7 +2,7 @@ import browserama from 'browserama';
 import { PureCloudWebrtcSdk } from './client';
 import { log } from './logging';
 import { LogLevels, SdkErrorTypes } from './types/enums';
-import { IMediaRequestOptions, IEnumeratedDevices, ISdkConfig, KeyFrom } from './types/interfaces';
+import { IMediaRequestOptions, IEnumeratedDevices } from './types/interfaces';
 import { throwSdkError } from './utils';
 
 const PC_AUDIO_EL_CLASS = '__pc-webrtc-inbound';
@@ -32,18 +32,9 @@ export const startDisplayMedia = function (): Promise<MediaStream> {
 export const startMedia = async function (sdk: PureCloudWebrtcSdk, opts: IMediaRequestOptions = { video: true, audio: true }): Promise<MediaStream> {
   const constraints: any = getStandardConstraints(opts);
 
-  /**
-   * Look for the desired deviceId
-   *  if found, use it
-   *    not found OR mediaType === true, find sdk default
-   *      if found, use it
-   *      not found, just use `true` for system default
-   */
-
   // if we are requesting video
-  if (opts.video) {
-    const videoDeviceId = await getDeviceByKindAndId(sdk, 'videoinput', opts.video);
-
+  if (opts.video || opts.video === null) {
+    const videoDeviceId = await getValidDeviceId(sdk, 'videoinput', opts.video);
     if (videoDeviceId) {
       log.call(sdk, LogLevels.info, 'Requesting video with deviceId', { deviceId: videoDeviceId });
       constraints.video.deviceId = {
@@ -55,8 +46,8 @@ export const startMedia = async function (sdk: PureCloudWebrtcSdk, opts: IMediaR
   }
 
   // if we are requesting audio
-  if (opts.audio) {
-    const audioDeviceId = await getDeviceByKindAndId(sdk, 'audioinput', opts.audio);
+  if (opts.audio || opts.audio === null) {
+    const audioDeviceId = await getValidDeviceId(sdk, 'audioinput', opts.audio);
 
     if (audioDeviceId) {
       log.call(sdk, LogLevels.info, 'Requesting audio with deviceId', { deviceId: audioDeviceId });
@@ -188,11 +179,11 @@ function getStandardConstraints (opts: IMediaRequestOptions): MediaStreamConstra
     };
   }
 
-  if (!opts.video) {
+  if (opts.video === false || opts.video === undefined) {
     constraints.video = false;
   }
 
-  if (!opts.audio) {
+  if (opts.audio === false || opts.audio === undefined) {
     constraints.audio = false;
   }
 
@@ -206,9 +197,12 @@ const enumeratedDevices: IEnumeratedDevices = {
   outputDeviceIds: []
 };
 
-async function getEnumeratedDevices (sdk: PureCloudWebrtcSdk): Promise<IEnumeratedDevices> {
+export async function getEnumeratedDevices (sdk: PureCloudWebrtcSdk): Promise<IEnumeratedDevices> {
   if (!window.navigator.mediaDevices || !window.navigator.mediaDevices.enumerateDevices) {
     log.call(sdk, LogLevels.warn, 'Unable to enumerate devices');
+    enumeratedDevices.videoDeviceIds = [];
+    enumeratedDevices.audioDeviceIds = [];
+    enumeratedDevices.outputDeviceIds = [];
     return enumeratedDevices;
   }
 
@@ -221,6 +215,7 @@ async function getEnumeratedDevices (sdk: PureCloudWebrtcSdk): Promise<IEnumerat
 
   // if devices haven't changed since last time we called this
   if (!refreshDevices) {
+    log.call(sdk, LogLevels.debug, 'Returning cached enumerated devices');
     return enumeratedDevices;
   }
 
@@ -236,7 +231,7 @@ async function getEnumeratedDevices (sdk: PureCloudWebrtcSdk): Promise<IEnumerat
         enumeratedDevices.videoDeviceIds.push(device.deviceId);
       } else if (device.kind === 'audioinput') {
         enumeratedDevices.audioDeviceIds.push(device.deviceId);
-      } else if (device.kind === 'audiooutput') {
+      } else /* if (device.kind === 'audiooutput') */ {
         enumeratedDevices.outputDeviceIds.push(device.deviceId);
       }
     });
@@ -247,11 +242,22 @@ async function getEnumeratedDevices (sdk: PureCloudWebrtcSdk): Promise<IEnumerat
   return enumeratedDevices;
 }
 
-// TODO: doc
-//  behavior, look for requested deviceId - if found return it
-//   if not found, look for the sdk default device id - if found return it
-//   if not found, return `undefined`
-export async function getDeviceByKindAndId (sdk: PureCloudWebrtcSdk, kind: MediaDeviceKind, deviceId: string | true): Promise<undefined | string> {
+/**
+ * Get a valid deviceId (`undefined` is returned if none was found)
+ *
+ * This will follow the steps looking for a device:
+ *  1. If `deviceId` is a `string`, it will look for that device and return it
+ *  2. If device could not be found _or_ `deviceId` was `true`,
+ *      it will look for the sdk default device
+ *  3. If device could not be found _or_ `deviceId` was `null`,
+ *      it will look for the system default device
+ *  4. If no device was found, return `undefined`
+ *
+ * @param sdk purecloud sdk instance
+ * @param kind desired device kind
+ * @param deviceId `deviceId` for specific device, `true` for sdk default device, or `null` for system default
+ */
+export async function getValidDeviceId (sdk: PureCloudWebrtcSdk, kind: MediaDeviceKind, deviceId: string | boolean | null): Promise<undefined | string> {
   const devices = await getEnumeratedDevices(sdk);
 
   let availableDevices: string[];
@@ -270,7 +276,9 @@ export async function getDeviceByKindAndId (sdk: PureCloudWebrtcSdk, kind: Media
   }
 
   // if a deviceId was passed in, try to use it
-  foundDeviceId = availableDevices.find((d: string) => d === deviceId);
+  if (typeof deviceId === 'string') {
+    foundDeviceId = availableDevices.find((d: string) => d === deviceId);
+  }
 
   // log if we didn't find the requested deviceId
   if (!foundDeviceId) {
@@ -284,6 +292,13 @@ export async function getDeviceByKindAndId (sdk: PureCloudWebrtcSdk, kind: Media
         log.call(sdk, LogLevels.warn, `Unable to find the sdk default ${kind} deviceId`, { deviceId: sdk._config.defaultAudioDeviceId });
       }
     }
+  }
+
+  /* if we are requesting 'audiooutput' and haven't found a device yet,
+    use the first output device in the list as it's the default. */
+  if (!foundDeviceId && kind === 'audiooutput') {
+    foundDeviceId = availableDevices[0];
+    log.call(sdk, LogLevels.info, `Using the system default 'audiooutput' device`, { deviceId: foundDeviceId });
   }
 
   return foundDeviceId;