+++ src/sessions/session-manager.ts
@@ -6,7 +6,7 @@ import { LogLevels, SessionTypes, SdkErrorTypes } from '../types/enums';
 import { throwSdkError } from '../utils';
 import ScreenShareSessionHandler from './screen-share-session-handler';
 import VideoSessionHandler from './video-session-handler';
-import { getDeviceByKindAndId } from '../media-utils';
+import { getValidDeviceId } from '../media-utils';
 import {
   IPendingSession,
   ISessionInfo,
@@ -111,12 +111,15 @@ export class SessionManager {
     return handler.startSession(startSessionParams);
   }
 
-  // TODO: doc
+  /**
+   * Update the outgoing media for a session.
+   *
+   * @param options for updating outgoing media
+   */
   async updateOutgoingMedia (options: IUpdateOutgoingMedia): Promise<any> {
     const session = options.session || this.getSession({ id: options.sessionId });
     const handler = this.getSessionHandler({ jingleSession: session });
 
-    // const stream = options.stream
     return handler.updateOutgoingMedia(session, options);
   }
 
@@ -132,20 +135,20 @@ export class SessionManager {
     return Promise.all(promises);
   }
 
-  async updateAudioOutputDeviceForAllSessions (outputDeviceId: string): Promise<any> {
-    const _outputDeviceId = await getDeviceByKindAndId(this.sdk, 'audiooutput', outputDeviceId);
+  async updateOutputDeviceForAllSessions (outputDeviceId: string | boolean | null): Promise<any> {
+    const _outputDeviceId = await getValidDeviceId(this.sdk, 'audiooutput', outputDeviceId);
 
     if (!_outputDeviceId) {
       this.log(LogLevels.warn, 'Output deviceId not found. Not updating output media', { outputDeviceId });
       return;
     }
 
-    const sessions = this.getAllActiveSessions();
-    this.log(LogLevels.info, 'Updating output deviceId for all active sessions', { sessions: sessions.map(s => s.id), outputDeviceId });
+    const sessions = this.getAllActiveSessions().filter(s => s.sessionType !== SessionTypes.acdScreenShare);
+    this.log(LogLevels.info, 'Updating output deviceId for all active sessions', { sessions: sessions.map(s => s.id), outputDeviceId: _outputDeviceId });
 
     const promises = sessions.map(session => {
       const handler = this.getSessionHandler({ jingleSession: session });
-      return handler.updateOutputDevice(session, outputDeviceId);
+      return handler.updateOutputDevice(session, _outputDeviceId);
     });
 
     return Promise.all(promises);