+++ src/sessions/video-session-handler.ts
@@ -1,7 +1,7 @@
 import BaseSessionHandler from './base-session-handler';
 import { IPendingSession, IStartSessionParams, IAcceptSessionRequest, ISessionMuteRequest, IJingleSession, IConversationUpdate, IParticipantUpdate, IParticipantsUpdate, IOnScreenParticipantsUpdate, ISpeakersUpdate, IConversationParticipant } from '../types/interfaces';
 import { SessionTypes, LogLevels, SdkErrorTypes, CommunicationStates } from '../types/enums';
-import { createNewStreamWithTrack, startMedia, startDisplayMedia } from '../media-utils';
+import { createNewStreamWithTrack, startMedia, startDisplayMedia, getValidDeviceId } from '../media-utils';
 import { throwSdkError, requestApi, isVideoJid } from '../utils';
 import { differenceBy, intersection } from 'lodash';
 
@@ -353,8 +353,8 @@ export default class VideoSessionHandler extends BaseSessionHandler {
 
       // if we are unmuting, we need to get a new camera track and add that to the session
     } else {
-      // if they passed in a deviceId, we will use that
-      const stream = await startMedia(this.sdk, { video: params.unmuteDeviceId || true });
+      // look for a device to use, else use default
+      const stream = await startMedia(this.sdk, { video: params.unmuteDeviceId === undefined ? true : params.unmuteDeviceId });
 
       // add track to session
       await this.addMediaToSession(session, stream, false);
@@ -387,7 +387,7 @@ export default class VideoSessionHandler extends BaseSessionHandler {
       track.enabled = !params.mute;
     });
 
-    if (params.mute && typeof params.mute !== 'string') {
+    if (params.mute) {
       if (!outgoingTracks.length) {
         this.log(LogLevels.warn, 'Unable to find any outgoing audio tracks to mute', { sessionId: session.id });
       } else {
@@ -398,8 +398,8 @@ export default class VideoSessionHandler extends BaseSessionHandler {
       if (!outgoingTracks.length) {
         this.log(LogLevels.info, 'No outoing audio to unmute, creating and adding media to session', { sessionId: session.id });
 
-        // if they passed in a deviceId, we will use that
-        const stream = await startMedia(this.sdk, { audio: params.unmuteDeviceId || true });
+        // if params.unmuteDeviceId is `undefined`, use sdk defaults
+        const stream = await startMedia(this.sdk, { audio: params.unmuteDeviceId === undefined ? true : params.unmuteDeviceId });
         await this.addMediaToSession(session, stream, false);
       }
 
@@ -407,6 +407,13 @@ export default class VideoSessionHandler extends BaseSessionHandler {
     }
 
     session.audioMuted = !!params.mute;
+
+    // if they passed in an unmute device id, we will switch to that device (if we unmuted audio)
+    if (params.unmuteDeviceId !== undefined && !session.audioMuted) {
+      this.log(LogLevels.info, 'switching audio device', { sessionId: session.id });
+
+      await this.sdk.updateOutgoingMedia({ audioDeviceId: params.unmuteDeviceId });
+    }
   }
 
   handleMediaChangeEvent (session: IJingleSession, event: IMediaChangeEvent) {