+++ changelog.md
@@ -1,14 +1,47 @@
-#Changelog
+# Changelog
+All notable changes to this project will be documented in this file.
 
-## v3.1.0
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+
+# [Unreleased]
+
+### Added
+* support for input and output devices
 * Screen pinning for video
-* Input and Output device management
 
-## v3.0.0
-* Change to a typescript implementation
-* Add `sdk.acceptSession` to be used in lieu of `session.accept()` in order to do some things before `session.accept()`
-* Add support for upcoming video conferencing feature
-* Use the latest streaming client which supports reallocating a new channel on no_longer_subscribed
+# [v3.1.0]
 
-#### Breaking Changes
+### Deprecated
+* SDK config option `iceTransportPolicy`
+
+### Added
+* Listeners to the video screen sharing stream to update the session when a user clicks the browser "Stop Sharing Screen" button
+
+### Changed
+* Do not request ice servers if there is no connection
+* Set iceTransportPolicy to `relay` if only turn servers are received
+* Update some typings
+
+# [v3.0.1]
+
+### Changed
+* Stop video conference screen share stream when the session ends
+
+# [v3.0.0]
+
+### Breaking Changes
 * Calling `session.accept()` manually which was the normal workflow after receiving the `sessionStarted` event is no longer supported and may have unpredictable outcomes. Use `sdk.acceptSession(options)` now.
+
+### Added
+* `sdk.acceptSession` to be used in lieu of `session.accept()` in order to do some things before `session.accept()`
+
+#### Changed
+* Moved code base to a typescript implementation
+
+
+[Unreleased]: https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.1.0...HEAD
+[v3.1.0]: https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.0.1...v3.1.0
+[v3.0.1]: https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.0.0...v3.0.1
+[v3.0.0]: https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v2.1.0...v3.0.0